+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed:
 - Habit completion percentages
+- Adaptive minY value in habit completion chart
 
 ## [0.8.242] - 2023-01-16
 ### Fixed: